Add hasCompletedDisarr function
