+++ packages/core/phase1/triggers/TRPS0010.ts
@@ -2,20 +2,11 @@ import TriggerCode from "bichard7-next-data-latest/dist/types/TriggerCode"
 import type { TriggerGenerator } from "../../phase1/types/TriggerGenerator"
 import isRecordableOffence from "../../phase2/isRecordableOffence"
 import disarrCompatibleResultClass from "../../phase2/lib/deriveOperationSequence/disarrCompatibleResultClass"
-import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import Phase from "../../types/Phase"
-import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
+import hasCompletedDisarr from "./hasCompletedDisarr"
 
 const triggerCode = TriggerCode.TRPS0010
-
-const hasCompletedDisarr = (pncUpdateDataset: PncUpdateDataset, offence: Offence) =>
-  pncUpdateDataset.PncOperations.some(
-    ({ code, status, data }) =>
-      code === "DISARR" &&
-      status === "Completed" &&
-      (!data?.courtCaseReference || data.courtCaseReference === offence.CourtCaseReferenceNumber)
-  )
 
 const generator: TriggerGenerator = (hearingOutcome, options) => {
   if (options?.phase !== Phase.PNC_UPDATE || !isPncUpdateDataset(hearingOutcome)) {
